+++ resources/js/components/GraphqlMutation.vue
@@ -52,7 +52,7 @@ export default {
         },
         store: {
             type: String,
-            default: window.config.store_code,
+            default: () => window.config.store_code,
         },
         beforeRequest: {
             type: Function,