+++ resources/js/components/Listing/Listing.vue
@@ -7,7 +7,7 @@ export default {
     props: {
         index: {
             type: String,
-            default: window.config.index.product,
+            default: () => window.config.index.product,
         },
         query: {
             type: Function,
@@ -119,10 +119,7 @@ export default {
                 return this.baseFilters().concat([
                     {
                         query_string: {
-                            query:
-                                '(visibility:(2 OR 4) OR (NOT _exists_:visibility)) AND (category_ids:' +
-                                this.categoryId +
-                                ' OR (NOT _exists_:category_ids))',
+                            query: `(visibility:(2 OR 4) OR (NOT _exists_:visibility)) AND (category_ids:${this.categoryId})`,
                         },
                     },
                 ])