+++ resources/js/components/Notifications/Notifications.vue
@@ -1,4 +1,7 @@
 <script>
+import { pendingNotifications, clearNotifications, notificationCount } from '../../stores/useNotifications.js'
+import { watch } from 'vue'
+
 import notification from './Notification.vue'
 Vue.component('notification', notification)
 
@@ -10,15 +13,24 @@ export default {
         return this.$scopedSlots.default(this)
     },
     mounted() {
-        this.$root.$on('notification-message', (message, type, params, link) => {
-            this.notifications.push({
-                message: message,
-                type: type,
-                params: params,
-                link: link,
-                show: true,
+        this.triggerNotifications()
+        watch(notificationCount, this.triggerNotifications)
+    },
+    methods: {
+        triggerNotifications() {
+            if (notificationCount.value == 0) {
+                return
+            }
+
+            pendingNotifications.value.forEach((notification) => {
+                this.notifications.push({
+                    show: true,
+                    ...notification,
+                })
             })
-        })
+
+            clearNotifications()
+        },
     },
 }
 </script>