+++ resources/views/layouts/partials/debug.blade.php
@@ -1,0 +1,7 @@
+<div
+    v-cloak
+    v-if="$root.configError"
+    class="w-full bg-danger text-center text-white font-bold"
+>
+    Config.js failed to load due to an error.
+</div>